requirements.txt
attrs==23.1.0
certifi==2023.7.22
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
PySocks==1.7.1
pytesseract==0.3.10
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
